requirements-gpu.txt
tensorflow-gpu=1.12
keras-gpu
numpy
pillow
matplotlib
scikit-learn